13/02/24

+++ Tema4/src/Ejercicio47/Articulo.java
@@ -1,16 +1,17 @@
 package Ejercicio47;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public abstract class Articulo {
 	
 	protected String descripcion;
-	protected Integer precio;
+	protected BigDecimal precio;
 	
-	public Integer getPrecio() {
+	public BigDecimal getPrecio() {
 		return precio;
 	}
-	public void setPrecio(Integer precio) {
+	public void setPrecio(BigDecimal precio) {
 		this.precio = precio;
 	}
